DAO reservation
methode findall
methode findById
ajout de constructeur par defaut sur plusieurs classes
supression de DaoFile (all)

+++ src/main/java/edu/projet/Interface.java
@@ -6,6 +6,6 @@ public interface Interface<T, PK>
 {
 	public List<T> findAll();
 
-	public T findById (Long id);
+	public T findById (int id);
 }
